+++ nutella/src/constant/Limit.ts
@@ -1,0 +1,3 @@
+const LIMIT = 20 as const;
+
+export default LIMIT;
