+++ nutella/src/interface/Pagination.ts
@@ -1,0 +1,6 @@
+interface Pagination {
+  limit: number;
+  page: number;
+}
+
+export default Pagination;
